More housekeeping (#23)
There is a small known issue that isn't harmful at all but just something to watch out for if we ever get confused about it. The TCP socket created by the memory viewer app only becomes non blocking after a successful connect, meaning that the reconnect attempt timeout also affects the windows polling of events, making it unresponsive in the meantime.

This of course means that if the engine isn't connected, the tool doesn't work properly, but for now I don't think it should be something to worry about. Still good to watch out for if we ever add saving and loading files to view inside the tool without the engine being open, etc.

+++ tools/memv/src/memv_app.cpp
@@ -161,8 +161,8 @@ namespace imp::tools::memv
             snap.peakUsed = alloc->peak_used();
             snap.totalAllocated = alloc->total_allocated();
             snap.totalFreed = alloc->total_freed();
-            snap.allocationCount = alloc->allocation_count();
-            snap.freeCount = alloc->free_count();
+            snap.allocationCount = static_cast<u32>( alloc->allocation_count() );
+            snap.freeCount = static_cast<u32>( alloc->free_count() );
 
             if (const auto *tagBytes = alloc->tag_bytes())
             {
